Moved default delays to index.ts
