+++ src/constants/index.ts
@@ -1,0 +1,3 @@
+export const DefaultMacroDelay = 20
+export const DefaultDelayDelay = 20
+export const DefaultMouseDelay = 20
