+++ src/constants/utils.ts
@@ -168,6 +168,4 @@ export const scrollbarsStylesDark = {
   }
 }
 
-export const DefaultMacroDelay = 20
-export const DefaultDelayDelay = 20
-export const DefaultMouseDelay = 20
+
